+++ dgl/alg/SplayTree.h
@@ -125,7 +125,7 @@ template<typename T>
 class SplayTree: public SearchTree<_node<T>> {
 public:
 	//Comp comp;
-	unsigned long p_size;
+	unsigned long long p_size;
 	typedef _node<T> Node;
 private:
 	//node * root;
@@ -374,7 +374,7 @@ public:
 	}
 	
 	//bool empty( ) const { return root == 0; }
-	unsigned long size() const {
+	unsigned long long size() const {
 		return p_size;
 	}
 };
